# Conflicts:
#	bookingpitch5/lib/api/create_pitch_service.dart
#	bookingpitch5/lib/view_models/my_booking_view_model.dart

+++ bookingpitch5/lib/models/pitchs/pitch_model.dart
@@ -7,7 +7,7 @@ class PitchModel {
   late String phone;
   late String time_start;
   late String time_end;
-  late int rates;
+  late double rates;
   late int status;
